feat(modularization): scaffold clawql-documents with external ingest (phase 8)
Introduce packages/clawql-documents for ingest_external_knowledge core logic
and URL formatting; keep MCP handler and mcp-tool-log in src/. Provider registry
remains in clawql-api (already extracted).
