+++ packages/clawql-documents/src/index.ts
@@ -1,0 +1,13 @@
+export {
+  externalIngestFeatureEnabled,
+  runIngestExternalKnowledge,
+  type ExternalIngestDocumentInput,
+  type ExternalIngestInput,
+  type ExternalIngestResult,
+} from "./ingest/external-ingest.js";
+export {
+  buildUrlIngestNote,
+  formatUrlResponseAsMarkdown,
+  type FormattedUrlIngest,
+  type UrlIngestKind,
+} from "./ingest/url-format.js";
